blog front posts

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -102,11 +102,12 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-        // $category = Category::find($id);
-        // $category->delete();
+        $category = Category::find($id);
+        if( $category->posts->count()){
+            return redirect()->route('categories.index')->with('error', 'У категории есть записи');
+        }
 
         Category::destroy($id);
-        
         return redirect()->route('categories.index')->with('success', 'Категория удалена');
     }
 }
