+++ app/Http/Controllers/Admin/TagController.php
@@ -99,6 +99,10 @@ class TagController extends Controller
      */
     public function destroy($id)
     {
+        $tag = Tag::find($id);
+        if($tag->posts->count()){
+            return redirect()->route('tags.index')->with('error', 'Ошибка у Tag есть записи');
+        }
        Tag::destroy($id);
        return redirect()->route('tags.index')->with('success', 'Tag удален');
     }
